re-activate EmbeddedTomcatWithSSLConfiguration (re. FINERACT-764)
It only didn't build anymore because the Commons IO was removed some time
(perhaps due to a version conflict?); but we don't really need it, it seems
to have been created for an obscure file path length problem on Windows?

If that Windows problem comes back, then we can always revert this, and properly fix.
(I may have actually coded that myself, many many years ago, it feels familiar.)
For now, this unblocks the FINERACT-764, req. for Gradle upgrade FINERACT-700.

Also delete EmbeddedTomcatWithSSLConfigurationTest, because it really
only tested that funky (and now deleted) resource extraction from JAR,
not really the EmbeddedTomcatWithSSLConfigurationTest.