remove Apache BVal @NotEmpty annotation use
I'm not sure if this even worked?  There's no test coverage for it.

This is required to be able to (easily) switch from openjpa-all to openjpa.

The alternative would be to add an explicit dependency to BVal;
but that seems unnecessary, and avoiding it will likely reduce other
dependency issues.